refactor: move cookie control JS files into civic_cookie_control/ folder
