+++ foundation_cms/static/js/civic_cookie_control/text.js
